+++ backend/car-service/src/main/java/com/rental/carservice/service/season/SeasonServiceImpl.java
@@ -1,11 +1,16 @@
 package com.rental.carservice.service.season;
 
-import com.rental.carservice.dto.SeasonDto;
+import com.rental.carservice.dto.season.SeasonCreationDto;
+import com.rental.carservice.dto.season.SeasonDto;
 import com.rental.carservice.mapper.SeasonMapper;
 import com.rental.carservice.model.Car;
+import com.rental.carservice.model.Group;
+import com.rental.carservice.model.Period;
 import com.rental.carservice.model.Season;
 import com.rental.carservice.repository.CarRepository;
+import com.rental.carservice.repository.GroupRepository;
 import com.rental.carservice.repository.SeasonRepository;
+import com.rental.carservice.util.DateConverter;
 import com.rental.carservice.util.ObjectValidation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -18,32 +23,25 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class SeasonServiceImpl implements SeasonService {
     private final SeasonRepository seasonRepository;
-    private final CarRepository carRepository;
+    private final GroupRepository groupRepository;
     private final SeasonMapper seasonMapper;
     private final ObjectValidation validation;
     @Override
     public List<SeasonDto> getAll(UUID companyId) {
-        List<Car> cars = carRepository.findAllByUserId(companyId);
-        Set<Season> seasons = new HashSet<>();
-
-        for (Car car : cars) {
-            seasons.addAll(car.getGroup().getSeasons());
-        }
-
-        return seasons.stream()
-                .map(seasonMapper::toDto)
+        return seasonRepository.findAllByCompanyId(companyId)
+                .stream().map(seasonMapper::toDto)
                 .collect(Collectors.toList());
     }
 
     @Override
-    public SeasonDto create(SeasonDto seasonDto) {
+    public SeasonDto create(SeasonCreationDto seasonDto) {
         Season season = new Season();
 
         season.setName(seasonDto.getName());
         season.setDescription(seasonDto.getDescription());
         season.setChangeValue(seasonDto.getChangeValue());
-        season.setStartDate(seasonDto.getStartDate());
-        season.setEndDate(seasonDto.getEndDate());
+        season.setStartDate(DateConverter.toDate(seasonDto.getStartDate()));
+        season.setEndDate(DateConverter.toDate(seasonDto.getEndDate()));
         season.setFixed(seasonDto.getIsFixed());
         season.setLastUpdated(OffsetDateTime.now());
 
@@ -51,13 +49,15 @@ public class SeasonServiceImpl implements SeasonService {
     }
 
     @Override
-    public SeasonDto edit(UUID id, SeasonDto seasonDto) {
+    public SeasonDto edit(UUID id, SeasonCreationDto seasonDto) {
         Optional<Season> seasonData = seasonRepository.findById(id).map(season -> {
                     String name = validation.setValue(seasonDto.getName(), season.getName());
                     String description = validation.setValue(seasonDto.getDescription(), season.getDescription());
                     int changeValue = validation.setValue(seasonDto.getChangeValue(), season.getChangeValue());
-                    OffsetDateTime startDate = validation.setValue(seasonDto.getStartDate(), season.getStartDate());
-                    OffsetDateTime endDate = validation.setValue(seasonDto.getEndDate(), season.getEndDate());
+                    OffsetDateTime startDate = validation
+                            .setValue(DateConverter.toDate(seasonDto.getStartDate()), season.getStartDate());
+                    OffsetDateTime endDate = validation
+                            .setValue(DateConverter.toDate(seasonDto.getEndDate()), season.getEndDate());
                     boolean isFixed = validation.setValue(seasonDto.getIsFixed(), season.isFixed());
 
                     season.setName(name);
@@ -68,14 +68,49 @@ public class SeasonServiceImpl implements SeasonService {
                     season.setFixed(isFixed);
                     season.setLastUpdated(OffsetDateTime.now());
 
-                    return season;
+                    return seasonRepository.save(season);
                 });
 
         return seasonData.map(seasonMapper::toDto).orElse(null);
     }
 
     @Override
-    public void delete(UUID id) {
+    public int addToGroup(UUID seasonId, UUID groupId) {
+        Season season = validation.getEntry(seasonRepository.findById(seasonId));
+        Group group = validation.getEntry(groupRepository.findById(groupId));
+        if (season != null && group != null) {
+            boolean isAdded = group.getSeasons().add(season);
+            if (!isAdded) {
+                return 409;
+            }
+            groupRepository.save(group);
+            return 204;
+        }
+        return 404;
+    }
+
+    @Override
+    public int removeFromGroup(UUID seasonId, UUID groupId) {
+        Season season = validation.getEntry(seasonRepository.findById(seasonId));
+        Group group = validation.getEntry(groupRepository.findById(groupId));
+        if (season != null && group != null) {
+            boolean isRemoved = group.getSeasons().remove(season);
+            if (!isRemoved) {
+                return 410;
+            }
+            groupRepository.save(group);
+            return 204;
+        }
+        return 404;
+    }
+
+    @Override
+    public int delete(UUID id) {
+        Season season = validation.getEntry(seasonRepository.findById(id));
+        if (season == null) {
+            return 404;
+        }
         seasonRepository.deleteById(id);
+        return seasonRepository.findById(id).isPresent() ? 500 : 204;
     }
 }
